requirements.txt
flask
psycopg2-binary
mysql-connector-python
flask-cors
pyjwt
python-dotenv